warnFix: BF and CBF explicit casts)

+++ BloomFilter.hpp
@@ -10,7 +10,7 @@ class BloomFilter {
 	 * @param num_hashes The number of hash function used to add a new value
 	 * @param reset_ratio The ratio of bits needed inside of the BF to invoque a self cleaning.
 	 */
-	BloomFilter(uint64_t size, uint8_t num_hashes, float reset_ratio);
+	BloomFilter(uint64_t size, uint8_t num_hashes, double reset_ratio);
 
 	void add(const uint8_t* data, std::size_t len);
 	uint64_t nbBitsSet() const;
@@ -24,5 +24,5 @@ class BloomFilter {
 	BitSet m_bits;
 	uint8_t m_num_hashes;
 	uint64_t m_bits_set;
-	float m_reset_ratio;
+	double m_reset_ratio;
 };
